pyproject.toml
[tool.poetry]
name = "soulsgen"
version = "0.1.0"
description = ""
authors = ["alexgQQ <me@alexgrand.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.10.2"
numpy = "^1.22.1"
pandas = "^1.4.0"
Expanda = "^1.3.1"
tokenizers = "0.9"
requests-oauthlib = "^1.3.1"
torchserve = "^0.5.2"
Cython = "^0.29.27"
torch-model-archiver = "^0.5.2"
regex = "^2022.1.18"
matplotlib = "^3.5.1"
tqdm = "^4.62.3"
scikit-learn = "^1.0.2"
seaborn = "^0.11.2"
Jinja2 = "^3.0.3"
spacy = "^3.2.3"
nltk = "^3.7"
click = "^8.0.4"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"
twitter = {path = "twitter-poster", develop = true}
docker = "^5.0.3"
google-cloud-storage = "^2.2.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pyinstrument = "^4.1.1"

[tool.poetry.scripts]
soulsgen = "soulsgen.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
